utils/db_connection.py
import mysql.connector
import os
from dotenv import load_dotenv

load_dotenv()

def get_connection():
    try:
        conn = mysql.connector.connect(
            host=os.getenv("DB_HOST"),
            user=os.getenv("DB_USER"),
            password=os.getenv("DB_PASSWORD"),
            database=os.getenv("DB_NAME"),
            port=int(os.getenv("DB_PORT", 3306))
        )
        return conn
    except Exception as e:
        print("DB Connection Error:", e)
        return None


def test_db_connection():
    conn = get_connection()
    if conn:
        conn.close()
        return True
    return False
