requirements.txt
streamlit
boto3
mysql-connector-python
python-dotenv
pandas
